pyproject.toml
[project]
name = "three-way-matcher"
version = "1.0.0"
description = "3-Way Matcher Agent — Supply Chain Document Reconciliation"
requires-python = ">=3.12"
dependencies = [
    "llama-index-workflows[server]>=2.2.0",
    "google-generativeai>=0.8",
    "python-dotenv>=1.0",
]

[tool.llamadeploy]
name = "three-way-matcher"

[tool.llamadeploy.workflows]
three_way_matcher = "workflow:workflow"

[tool.llamadeploy.ui]
directory = "./ui"
